+++ local-data-platform/local_data_platform/pipeline/ingestion/__init__.py
@@ -1,11 +1,16 @@
 from local_data_platform.pipeline import Pipeline
+from local_data_platform.logger import log
+
+logger = log()
 
 
 class Ingestion(Pipeline):
 
-
     def extract(self):
-        self.source.get()
+        logger.info("Extracting Source in ingestion pipeline")
+        return self.source.get()
 
     def load(self):
-        self.target.put(self.extract())
+        df = self.extract()
+        logger.info(f"Loading Source {len(df)} in ingestion pipeline")
+        self.target.put(df)
