+++ local-data-platform/real_world_use_cases/nyc_yellow_taxi_dataset/config/egression.json
@@ -11,15 +11,15 @@
     "target": {
       "name": "rides",
       "format": "CSV",
-      "path": "/Users/tushar/Documents/GitHub/local-data-platform/local-data-platform/yellow_tripdata_2023-01.csv"
+      "path": "/real_world_use_cases/nyc_yellow_taxi_dataset/data/nyc_yellow_taxi_rides.csv"
     },
     "source": {
       "name": "rides",
       "format": "ICEBERG",
       "catalog": {
         "type": "LocalIceberg",
-        "identifier": "pyiceberg_catalog_db",
-        "warehouse_path": "./tmp/warehouse"
+        "identifier": "nyc_yellow_taxi_dataset",
+        "warehouse_path": "real_world_use_cases/nyc_yellow_taxi_dataset/data"
       }
     }
   }
