+++ local-data-platform/real_world_use_cases/nyc_yellow_taxi_dataset/config/ingestion.json
@@ -11,7 +11,7 @@
     "source": {
       "name": "rides",
       "format": "PARQUET",
-      "path": "/Users/tushar/Documents/GitHub/local-data-platform/local-data-platform/yellow_tripdata_2023-01.parquet"
+      "path": "/yellow_tripdata_2023-01.parquet"
     },
     "target": {
       "name": "rides",
@@ -19,8 +19,8 @@
       "path": "yellow_tripdata_2023-01.parquet",
       "catalog": {
         "type": "LocalIceberg",
-        "identifier": "pyiceberg_catalog_db",
-        "warehouse_path": "./tmp/warehouse"
+        "identifier": "nyc_yellow_taxi_dataset",
+        "warehouse_path": "real_world_use_cases/nyc_yellow_taxi_dataset/data"
       }
     }
   }
